requirements.txt
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
tqdm>=4.66.0
rouge-score>=0.1.2
bert-score>=0.3.13
sentence-transformers>=3.0.0
requests>=2.31.0
pypdf>=4.0.0
